[ INFO ] [warm-up][P0] Generated: <s> What is OpenVINO?
OpenVINO is a free downloadable software that uses the power of artificial machine learning to transform your business. It is a powerful combination of a processor and a neural network.

The main advantage of the program is that it is a cross-industry solution. It is equally well suited for both

[ INFO ] [warm-up][P0] Generated: <s> Who is the most famous mathematician?

The most famous mathematician in the 20th century is widely considered to be John von Neumann (1927-1995). He was a brilliant man who worked on the first computers and was one

Who is the most famous female mathemics?